Added standard polymorphic_value, added tracking structure

+++ include/utils/math/constants.h
@@ -8,6 +8,6 @@ namespace utils::constants
 	inline constexpr float PI = PIf;
 
 	inline constexpr float epsilonf = std::numeric_limits<float>::epsilon();
-	inline constexpr float epsilond = std::numeric_limits<double>::epsilon();
+	inline constexpr double epsilond = std::numeric_limits<double>::epsilon();
 	inline constexpr float epsilon = epsilonf;
 	}
